Rewrite GG decoder and add systematic scheduling
The GG decoder is rewritten to make use of the row_vector data
structure. This helps to save memory usage and would make the decoder
simpler.

Systematic scheduling is added to sncRecoder. For systematic codes, the
scheduling would forward each received systematic packet (i.e., if not
erased in the upstream hop). In line networks, this not only reduces
decoding cost, but also improves end-to-end rate because each systematic
packet is innovative for downstream nodes of the line network with
probability one.